+++ src/functions/color/hsl.rs
@@ -1,12 +1,12 @@
 use super::rgb::{preserve_call, values_from_list};
-use super::{get_color, make_call, Error, Module, SassFunction};
+use super::{get_color, make_call, Error, FunctionMap, SassFunction};
 use crate::css::Value;
 use crate::value::{Hsla, Numeric, Unit};
-use crate::variablescope::Scope;
+use crate::Scope;
 use num_rational::Rational;
 use num_traits::{one, One, Zero};
 
-fn do_hsla(fn_name: &str, s: &dyn Scope) -> Result<Value, Error> {
+fn do_hsla(fn_name: &str, s: &Scope) -> Result<Value, Error> {
     let a = s.get("alpha")?;
     let hue = s.get("hue")?;
     if let Value::List(vec, sep, bracketed) = if hue.is_null() {
@@ -45,14 +45,14 @@ fn hsla_from_values(
     Some(Hsla::new(h, s, l, a).into())
 }
 
-pub fn register(f: &mut Module) {
+pub fn register(f: &mut Scope) {
     def!(f, _hsl(hue, saturation, lightness, alpha, channels), |s| {
         do_hsla("hsl", s)
     });
     def!(f, _hsla(hue, saturation, lightness, alpha, channels), |s| {
         do_hsla("hsla", s)
     });
-    def!(f, _adjust_hue(color, degrees), |s: &dyn Scope| match (
+    def!(f, _adjust_hue(color, degrees), |s| match (
         s.get("color")?,
         s.get("degrees")?,
     ) {
@@ -65,7 +65,7 @@ pub fn register(f: &mut Module) {
     def!(f, complement(color), |s| {
         Ok(get_color(s, "color")?.rotate_hue(180.into()).into())
     });
-    def!(f, _saturate(color, amount), |s: &dyn Scope| match (
+    def!(f, _saturate(color, amount), |s| match (
         s.get("color")?,
         s.get("amount")?
     ) {
@@ -123,22 +123,22 @@ pub fn register(f: &mut Module) {
     });
 }
 
-pub fn expose(m: &Module, global: &mut Module) {
-    for &(gname, lname) in &[
-        ("hsl", "_hsl"),
-        ("hsla", "_hsla"),
-        ("adjust_hue", "_adjust_hue"),
-        ("complement", "complement"),
-        ("darken", "_darken"),
-        ("desaturate", "_desaturate"),
-        ("grayscale", "grayscale"),
-        ("hue", "hue"),
-        ("lighten", "_lighten"),
-        ("lightness", "lightness"),
-        ("saturate", "_saturate"),
-        ("saturation", "saturation"),
+pub fn expose(m: &Scope, global: &mut FunctionMap) {
+    for (gname, lname) in &[
+        (name!(hsl), name!(_hsl)),
+        (name!(hsla), name!(_hsla)),
+        (name!(adjust_hue), name!(_adjust_hue)),
+        (name!(complement), name!(complement)),
+        (name!(darken), name!(_darken)),
+        (name!(desaturate), name!(_desaturate)),
+        (name!(grayscale), name!(grayscale)),
+        (name!(hue), name!(hue)),
+        (name!(lighten), name!(_lighten)),
+        (name!(lightness), name!(lightness)),
+        (name!(saturate), name!(_saturate)),
+        (name!(saturation), name!(saturation)),
     ] {
-        global.expose(gname, m, lname);
+        global.insert(gname.clone(), m.get_function(&lname).unwrap().clone());
     }
 }